Crear compuertas AND moviles

+++ src/AbstractFactory/AbsCompuertaFactory.java
@@ -3,12 +3,12 @@ package AbstractFactory;
 import circuitDesing.Compuerta;
 
 public interface AbsCompuertaFactory {
-    public Compuerta crearAND(int X, int Y , int entradas, int ID);
-    public Compuerta crearNAND(int X, int Y , int entradas,int ID);
-    public Compuerta crearOR(int X, int Y , int entradas, int ID);
-    public Compuerta crearNOR(int X, int Y , int entradas, int ID);
-    public Compuerta crearNOT(int X, int Y, int ID);
-    public Compuerta crearXOR(int X, int Y , int entradas, int ID);
-    public Compuerta crearXNOR(int X, int Y , int entradas, int ID);
-    public Compuerta crearCustomGate(int X, int Y , int entradas, int salidas);
+    public Compuerta crearAND(double X, double Y , int entradas, int ID);
+    public Compuerta crearNAND(double X, double Y , int entradas,int ID);
+    public Compuerta crearOR(double X, double Y , int entradas, int ID);
+    public Compuerta crearNOR(double X, double Y , int entradas, int ID);
+    public Compuerta crearNOT(double X, double Y, int ID);
+    public Compuerta crearXOR(double X, double Y , int entradas, int ID);
+    public Compuerta crearXNOR(double X, double Y , int entradas, int ID);
+    public Compuerta crearCustomGate(double X, double Y , int entradas, int salidas);
 }
